chore: Update font tests & calculateDynamicFontSize
- Updated the font tests and added tests for calculateDynamicFontSize.
- Also refactored code for readability and maintainability.
