+++ packages/common/__tests__/helper.test.ts
@@ -1,185 +1,4 @@
-import { readFileSync } from 'fs';
-import * as path from 'path';
-import { Font, Template } from '../src/type';
-import { checkFont, validateBarcodeInput } from '../src/helper';
-import { BLANK_PDF } from '../src';
-
-const sansData = readFileSync(path.join(__dirname, `/assets/fonts/SauceHanSansJP.ttf`));
-const serifData = readFileSync(path.join(__dirname, `/assets/fonts/SauceHanSerifJP.ttf`));
-
-const getSampleFont = (): Font => ({
-  SauceHanSansJP: { fallback: true, data: sansData },
-  SauceHanSerifJP: { data: serifData },
-});
-
-const getTemplate = (): Template => ({
-  basePdf: BLANK_PDF,
-  schemas: [
-    {
-      a: {
-        type: 'text',
-        fontName: 'SauceHanSansJP',
-        position: { x: 0, y: 0 },
-        width: 100,
-        height: 100,
-      },
-      b: {
-        type: 'text',
-        position: { x: 0, y: 0 },
-        width: 100,
-        height: 100,
-      },
-    },
-  ],
-});
-
-describe('checkFont test', () => {
-  test('success test: no fontName in Schemas', () => {
-    const _getTemplate = (): Template => ({
-      basePdf: BLANK_PDF,
-      schemas: [
-        {
-          a: {
-            type: 'text',
-            position: { x: 0, y: 0 },
-            width: 100,
-            height: 100,
-          },
-          b: {
-            type: 'text',
-            position: { x: 0, y: 0 },
-            width: 100,
-            height: 100,
-          },
-        },
-      ],
-    });
-    try {
-      checkFont({ template: _getTemplate(), font: getSampleFont() });
-      expect.anything();
-    } catch (e) {
-      fail();
-    }
-  });
-
-  test('success test: fontName in Schemas(fallback font)', () => {
-    try {
-      checkFont({ template: getTemplate(), font: getSampleFont() });
-      expect.anything();
-    } catch (e) {
-      fail();
-    }
-  });
-
-  test('success test: fontName in Schemas(not fallback font)', () => {
-    const getFont = (): Font => ({
-      SauceHanSansJP: { data: sansData },
-      SauceHanSerifJP: { fallback: true, data: serifData },
-    });
-
-    try {
-      checkFont({ template: getTemplate(), font: getFont() });
-      expect.anything();
-    } catch (e) {
-      fail();
-    }
-  });
-
-  test('fail test: no fallback font', () => {
-    const getFont = (): Font => ({
-      SauceHanSansJP: { data: sansData },
-      SauceHanSerifJP: { data: serifData },
-    });
-
-    try {
-      checkFont({ template: getTemplate(), font: getFont() });
-      fail();
-    } catch (e: any) {
-      expect(e.message).toEqual(
-        'fallback flag is not found in font. true fallback flag must be only one.'
-      );
-    }
-  });
-
-  test('fail test: too many fallback font', () => {
-    const getFont = (): Font => ({
-      SauceHanSansJP: { data: sansData, fallback: true },
-      SauceHanSerifJP: { data: serifData, fallback: true },
-    });
-
-    try {
-      checkFont({ template: getTemplate(), font: getFont() });
-      fail();
-    } catch (e: any) {
-      expect(e.message).toEqual(
-        '2 fallback flags found in font. true fallback flag must be only one.'
-      );
-    }
-  });
-
-  test('fail test: fontName in Schemas not found in font(single)', () => {
-    const _getTemplate = (): Template => ({
-      basePdf: BLANK_PDF,
-      schemas: [
-        {
-          a: {
-            type: 'text',
-            fontName: 'SauceHanSansJP2',
-            position: { x: 0, y: 0 },
-            width: 100,
-            height: 100,
-          },
-          b: {
-            type: 'text',
-            position: { x: 0, y: 0 },
-            width: 100,
-            height: 100,
-          },
-        },
-      ],
-    });
-
-    try {
-      checkFont({ template: _getTemplate(), font: getSampleFont() });
-      fail();
-    } catch (e: any) {
-      expect(e.message).toEqual('SauceHanSansJP2 of template.schemas is not found in font.');
-    }
-  });
-
-  test('fail test: fontName in Schemas not found in font(single)', () => {
-    const _getTemplate = (): Template => ({
-      basePdf: BLANK_PDF,
-      schemas: [
-        {
-          a: {
-            type: 'text',
-            fontName: 'SauceHanSansJP2',
-            position: { x: 0, y: 0 },
-            width: 100,
-            height: 100,
-          },
-          b: {
-            type: 'text',
-            fontName: 'SauceHanSerifJP2',
-            position: { x: 0, y: 0 },
-            width: 100,
-            height: 100,
-          },
-        },
-      ],
-    });
-
-    try {
-      checkFont({ template: _getTemplate(), font: getSampleFont() });
-      fail();
-    } catch (e: any) {
-      expect(e.message).toEqual(
-        'SauceHanSansJP2,SauceHanSerifJP2 of template.schemas is not found in font.'
-      );
-    }
-  });
-});
+import { validateBarcodeInput } from '../src/helper';
 
 describe('validateBarcodeInput test', () => {
   test('qrcode', () => {
