+++ packages/common/src/helper.ts
@@ -1,7 +1,6 @@
 import { z } from 'zod';
 import { Buffer } from 'buffer';
-import { DEFAULT_FONT_NAME, DEFAULT_FONT_VALUE } from './constants.js';
-import { Template, Schema, BasePdf, Font, CommonProps, isTextSchema, BarCodeType } from './type.js';
+import { BasePdf, CommonProps, BarCodeType, } from './type';
 import {
   Inputs as InputsSchema,
   UIOptions as UIOptionsSchema,
@@ -10,8 +9,8 @@ import {
   DesignerProps as DesignerPropsSchema,
   GenerateProps as GeneratePropsSchema,
   UIProps as UIPropsSchema,
-} from './schema.js';
-import type { Font as FontkitFont } from 'fontkit';
+} from './schema';
+import { checkFont } from "./font"
 
 const blob2Base64Pdf = (blob: Blob) => {
   return new Promise<string>((resolve, reject) => {
@@ -54,74 +53,6 @@ export const b64toUint8Array = (base64: string) => {
     unit8arr[i] = byteString.charCodeAt(i);
   }
   return unit8arr;
-};
-
-export const getFallbackFontName = (font: Font) => {
-  const initial = '';
-  const fallbackFontName = Object.entries(font).reduce((acc, cur) => {
-    const [fontName, fontValue] = cur;
-
-    return !acc && fontValue.fallback ? fontName : acc;
-  }, initial);
-  if (fallbackFontName === initial) {
-    throw Error(`fallback flag is not found in font. true fallback flag must be only one.`);
-  }
-
-  return fallbackFontName;
-};
-
-export const getDefaultFont = (): Font => ({
-  [DEFAULT_FONT_NAME]: { data: b64toUint8Array(DEFAULT_FONT_VALUE), fallback: true },
-});
-
-export const heightOfFontAtSize = (font: FontkitFont, size: number) => {
-  let { ascent, descent, bbox } = font;
-
-  const scale = 1000 / font.unitsPerEm;
-  const yTop = (ascent || bbox.maxY) * scale;
-  const yBottom = (descent || bbox.minY) * scale;
-
-  let height = yTop - yBottom;
-  height -= Math.abs(descent * scale) || 0;
-
-  return (height / 1000) * size;
-};
-
-const uniq = <T>(array: Array<T>) => Array.from(new Set(array));
-
-const getFontNamesInSchemas = (schemas: { [key: string]: Schema }[]) =>
-  uniq(
-    schemas
-      .map((s) => Object.values(s).map((v) => (isTextSchema(v) ? v.fontName : '')))
-      .reduce((acc, cur) => acc.concat(cur), [] as (string | undefined)[])
-      .filter(Boolean) as string[]
-  );
-
-export const checkFont = (arg: { font: Font; template: Template }) => {
-  const {
-    font,
-    template: { schemas },
-  } = arg;
-  const fontValues = Object.values(font);
-  const fallbackFontNum = fontValues.reduce((acc, cur) => (cur.fallback ? acc + 1 : acc), 0);
-  if (fallbackFontNum === 0) {
-    throw Error(`fallback flag is not found in font. true fallback flag must be only one.`);
-  }
-  if (fallbackFontNum > 1) {
-    throw Error(
-      `${fallbackFontNum} fallback flags found in font. true fallback flag must be only one.`
-    );
-  }
-
-  const fontNamesInSchemas = getFontNamesInSchemas(schemas);
-  const fontNames = Object.keys(font);
-  if (fontNamesInSchemas.some((f) => !fontNames.includes(f))) {
-    throw Error(
-      `${fontNamesInSchemas
-        .filter((f) => !fontNames.includes(f))
-        .join()} of template.schemas is not found in font.`
-    );
-  }
 };
 
 const checkProps = <T>(data: unknown, zodSchema: z.ZodType<T>) => {
